requirements.txt
MouseInfo==0.1.3
numpy==2.3.4
pillow==12.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
python3-xlib==0.15
pytweening==1.2.0
